H2 Completado

+++ FP_T2/src/H2/UltraBall.java
@@ -9,7 +9,7 @@ public class UltraBall extends Pokeball
 	{
 		return potencia;
 	}
-	public void setPotencia(int potencia) 
+	public void setPotencia(double potencia) 
 	{
 		this.potencia = potencia;
 	}
